rename sky -> flutter in shell
