+++ shell/platform/darwin/desktop/flutter_app_delegate.h
@@ -1,0 +1,14 @@
+// Copyright 2015 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#ifndef __SHELL_PLATFORM_DARWIN_DESKTOP_FLUTTER_APP_DELEGATE__
+#define __SHELL_PLATFORM_DARWIN_DESKTOP_FLUTTER_APP_DELEGATE__
+
+#import <Cocoa/Cocoa.h>
+
+@interface FlutterAppDelegate : NSObject<NSApplicationDelegate>
+
+@end
+
+#endif /* defined(__SHELL_PLATFORM_DARWIN_DESKTOP_FLUTTER_APP_DELEGATE__) */
