+++ shell/platform/darwin/desktop/flutter_app_delegate.m
@@ -1,0 +1,15 @@
+// Copyright 2015 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#import "flutter_app_delegate.h"
+
+@interface FlutterAppDelegate ()
+
+@property(assign) IBOutlet NSWindow* window;
+
+@end
+
+@implementation FlutterAppDelegate
+
+@end
