--- shell/platform/darwin/desktop/sky_app_delegate.h
@@ -1,9 +1,0 @@
-// Copyright 2015 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-#import <Cocoa/Cocoa.h>
-
-@interface SkyAppDelegate : NSObject<NSApplicationDelegate>
-
-@end
