--- shell/platform/darwin/desktop/sky_app_delegate.m
@@ -1,15 +1,0 @@
-// Copyright 2015 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-#import "sky_app_delegate.h"
-
-@interface SkyAppDelegate ()
-
-@property(assign) IBOutlet NSWindow* window;
-
-@end
-
-@implementation SkyAppDelegate
-
-@end
